backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
python-dotenv
psycopg2-binary
requests
supabase
SQLAlchemy
